Changed OpenEndedAwnser
Now the DAO class throws exceptions like it suposed.
All the long types were changed to Long.
The Management class is on the standard form.

+++ JQuestBusiness/src/br/cefetmg/jquest/model/service/OpenEndedManagementImpl.java
@@ -33,15 +33,6 @@ public class OpenEndedManagementImpl implements OpenEndedAnswerManagement{
     }
     
     /**
-     * Get the value of DAO
-     *
-     * @return the value of DAO
-     */
-    public OpenEndedAnswerDAO getDAO() {
-        return DAO;
-    }
-
-    /**
      * Set the value of DAO
      *
      * @param DAO new value of DAO
@@ -51,48 +42,64 @@ public class OpenEndedManagementImpl implements OpenEndedAnswerManagement{
     }
 
     @Override
-    public void OpenEndedAnswerInsert(OpenEndedAnswer openEndedAnswer) throws BusinessException, PersistenceException {
+    public Long OpenEndedAnswerInsert(OpenEndedAnswer openEndedAnswer) throws BusinessException, PersistenceException {
         if(openEndedAnswer == null){
-            throw new BusinessException("The object cannot be null.");
+            throw new BusinessException("The object OpenEndedAnswer cannot be null.");
+        }
+        if(openEndedAnswer.getIDQuestion() == null || openEndedAnswer.getIdUser() == null || openEndedAnswer.getSeqAnswerUser() == null){
+            throw new BusinessException("None of the idQuestion or iduser or seqAnswer can be null.");
         }
         if(openEndedAnswer.equals(new OpenEndedAnswer())){
-            throw new BusinessException("The object cannot be empty.");
+            throw new BusinessException("The object OpenEndedAnswer cannot be empty.");
         }
-        if(openEndedAnswer.getTxtAnswer().equals("")){
-            throw new BusinessException("The Answer text cannot be null.");
+        if(openEndedAnswer.getTxtAnswer().isEmpty()){
+            throw new BusinessException("The Answer text cannot be empty.");
         }
         
         DAO.insert(openEndedAnswer);
+        return openEndedAnswer.getSeqAnswerUser();
     }
 
     @Override
     public void OpenEndedAnswerUpdate(OpenEndedAnswer openEndedAnswer) throws BusinessException, PersistenceException {
-         if(openEndedAnswer == null){
+        if(openEndedAnswer == null){
             throw new BusinessException("The object cannot be null.");
+        }
+        if(openEndedAnswer.getIDQuestion() == null || openEndedAnswer.getIdUser() == null || openEndedAnswer.getSeqAnswerUser() == null){
+            throw new BusinessException("None of the idQuestion or iduser or seqAnswer can be null.");
         }
         if(openEndedAnswer.equals(new OpenEndedAnswer())){
             throw new BusinessException("The object cannot be empty.");
         }
-        if(openEndedAnswer.getTxtAnswer().equals("")){
-            throw new BusinessException("The Answer text cannot be null.");
+        if(openEndedAnswer.getTxtAnswer().isEmpty()){
+            throw new BusinessException("The Answer text cannot be empty.");
         }
         
         DAO.update(openEndedAnswer);
     }
 
     @Override
-    public void OpenEndedAnswerRemove(long idQuestion, long idUser, long seqAnswerUser) throws PersistenceException {
-         DAO.remove(idQuestion, idUser, seqAnswerUser);
+    public void OpenEndedAnswerRemove(Long idQuestion, Long idUser, Long seqAnswerUser) throws PersistenceException {
+        if(idQuestion == null || idUser == null || seqAnswerUser == null){
+            throw new PersistenceException("None of the parameters can be null.");
+        } 
+        DAO.remove(idQuestion, idUser, seqAnswerUser);
     }
 
     @Override
-    public OpenEndedAnswer getOpenEndedAnswerById(long idQuestion, long idUser, long seqAnswerUser) throws PersistenceException {
+    public OpenEndedAnswer getOpenEndedAnswerById(Long idQuestion, Long idUser, Long seqAnswerUser) throws PersistenceException {
+        if(idQuestion == null || idUser == null || seqAnswerUser == null){
+            throw new PersistenceException("None of the parameters can be null.");
+        }
         return DAO.getOpenEndedAnswerById(idQuestion, idUser, seqAnswerUser);
     }
 
     @Override
     public List<OpenEndedAnswer> getAll() throws PersistenceException {
-        return DAO.listAll();
+        List<OpenEndedAnswer> aux = DAO.listAll();
+        if(aux.isEmpty())
+            throw new PersistenceException("There isn't elements in the List.");
+        return aux;
     }
